move package in groovy files too
